# opentracing: Use protobufs for binary enc/decoding