--- a/opentracing/pb/propagation.proto
+++ b/opentracing/pb/propagation.proto
@@ -1,10 +1,10 @@
 syntax = "proto3";
 package pb;
 
-message TraceState {
+message TracerState {
   uint64 trace_id = 1;
   uint64 span_id = 3;
-  bool sampled_status = 2;
+  bool sampled = 2;
 }
 
 message Baggage {