Use protocol buffers instead of `encoding/binary`. Benchmarks run on an aws m3.large
```
benchmark                                  old ns/op     new ns/op     delta
BenchmarkBinaryJoin_0BaggageItems          1527          1421          -6.94%
BenchmarkBinaryJoin_100BaggageItems        150660        99847         -33.73%
BenchmarkBinaryJoin_1000BaggageItems       1521358       1023776       -32.71%
BenchmarkBinaryInject_0BaggageItems        1252          1079          -13.82%
BenchmarkBinaryInject_100BaggageItems      72363         39975         -44.76%
BenchmarkBinaryInject_1000BaggageItems     698984        387390        -44.58%

benchmark                                  old allocs     new allocs     delta
BenchmarkBinaryJoin_0BaggageItems          14             6              -57.14%
BenchmarkBinaryJoin_100BaggageItems        1421           620            -56.37%
BenchmarkBinaryJoin_1000BaggageItems       14056          6089           -56.68%
BenchmarkBinaryInject_0BaggageItems        9              8              -11.11%
BenchmarkBinaryInject_100BaggageItems      814            126            -84.52%
BenchmarkBinaryInject_1000BaggageItems     8017           1037           -87.06%

benchmark                                  old bytes     new bytes     delta
BenchmarkBinaryJoin_0BaggageItems          457           538           +17.72%
BenchmarkBinaryJoin_100BaggageItems        30224         18044         -40.30%
BenchmarkBinaryJoin_1000BaggageItems       317720        227992        -28.24%
BenchmarkBinaryInject_0BaggageItems        405           324           -20.00%
BenchmarkBinaryInject_100BaggageItems      16064         7578          -52.83%
BenchmarkBinaryInject_1000BaggageItems     147584        99891         -32.32%
```